// AngularMain/src/app/modules/models/ReferenceDetails/RefTableDto.ts
export interface RefTableDto {
        tenentId: number;
        refid: number;
        reftype: string;
        refsubtype: string;
        shortname: string;
        refname1: string;
        refname2: string;
        refname3: string;
        switch1: string;
        switch2: string;
        switch3: string;
        switch4: number;
        remarks: string;
        active: string;
        crupId: number;
        infrastructure: string;
        refImage: string;
        uploadDate: Date;
        uploadby: string;
        syncDate: Date;
        syncby: string;
        synId: number;
    }